Move operators to `operators` package
