+++ pkg/streamingpromql/operators/time.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package operator
+package operators
 
 func stepCount(start, end, interval int64) int {
 	return int((end-start)/interval) + 1
